Refactor: Extract Dates class into dedicated module
- Move Dates class from defs.py to defs/dates.py module
- Update imports
- Remove duplicated Dates implementation from defs.py

This allows Dates module to be reused in other scripts.
