+++ src/defs/defs.py
@@ -9,6 +9,7 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from types import ModuleType
 from typing import Dict, List, Optional, Tuple, Type, Union
+from .dates import Dates
 import itertools
 import dateutil
 
@@ -125,46 +126,6 @@ def load_module(module_str: str) -> Union[ModuleType, Type]:
     spec.loader.exec_module(module)
 
     return getattr(module, class_name) if class_name else module
-
-
-class Dates:
-    "A class for date related functions in EOD2"
-
-    def __init__(self, lastUpdate: str):
-        today = datetime.now(tz_IN)
-
-        self.today = datetime.combine(today, datetime.min.time())
-
-        dt = datetime.fromisoformat(lastUpdate).astimezone(tz_IN)
-
-        self.dt = self.lastUpdate = dt
-
-        self.pandasDt = self.dt.strftime("%Y-%m-%d")
-
-    def nextDate(self):
-        """Set the next trading date and return True.
-        If its a future date, return False"""
-
-        curTime = datetime.now(tz_IN)
-        self.dt = self.dt + timedelta(1)
-
-        if self.dt > curTime:
-            logger.info("All Up To Date")
-            return False
-
-        if self.dt.day == curTime.day and curTime.hour < 16:
-            # Display the users local time
-            local_time = curTime.replace(hour=19, minute=0).astimezone(tz_local)
-
-            t_str = local_time.strftime("%I:%M%p")  # 07:00PM
-
-            logger.info(
-                f"All Up To Date. Check again after {t_str} for today's EOD data"
-            )
-            return False
-
-        self.pandasDt = self.dt.strftime("%Y-%m-%d")
-        return True
 
 
 def log_unhandled_exception(exc_type, exc_value, exc_traceback):
